refactor the classes
