--- icu4j/main/classes/core/src/com/ibm/icu/impl/units/Convertibility.java
@@ -1,7 +1,0 @@
-package com.ibm.icu.impl.units;
-
-public enum Convertibility {
-    CONVERTIBLE,
-    RECIPROCAL,
-    INCONVERTIBLE,
-}
