+++ icu4j/main/classes/core/src/com/ibm/icu/impl/units/UnitConverter.java
@@ -90,6 +90,11 @@ public class UnitConverter {
         return true;
     }
 
+    public enum Convertibility {
+        CONVERTIBLE,
+        RECIPROCAL,
+        INCONVERTIBLE,
+    }
 }
